fix:review structure app address

+++ src/main/kotlin/server/web/casa/app/address/domain/model/District.kt
@@ -1,7 +1,7 @@
 package server.web.casa.app.address.domain.model
 
 data class District(
-    val districtId : Int,
+    val districtId : Long,
     val city   : City,
     val name   : String,
 //    val commune: Commune,
