+++ src/main/kotlin/server/web/casa/app/address/domain/model/Quartier.kt
@@ -1,7 +1,7 @@
 package server.web.casa.app.address.domain.model
 
 data class Quartier(
-    val quartierId  : Int,
+    val quartierId  : Long,
     val commune   : Commune,
     val name        : String,
 )
